add:controller tests

+++ backend/src/incomes/income.controller.ts
@@ -2,6 +2,8 @@ import express from "express";
 
 import Controller from "../common/controller";
 import IncomeNotFoundException from "../exceptions/IncomeNotFoundException";
+import NotAuthorizedException from "../exceptions/NotAuthorizedException";
+import RequestWithUser from "../interfaces/requestWithUser.interface";
 import authMiddleware from "../middleware/auth.middleware";
 import validationMiddleware from "../middleware/validation.middleware";
 import CreateIncomeDto from "./income.dto";
@@ -18,28 +20,39 @@ class IncomeController implements Controller {
   }
 
   private initializeRoutes() {
-    this.router.get(this.path, this.getAllIncomes);
-    this.router.get(`${this.path}/:id`, this.getIncomeById);
+    this.router.get(this.path, authMiddleware, this.getAllIncomes);
+    this.router.get(`${this.path}/:id`, authMiddleware, this.getIncomeById);
 
     this.router
       .all(`${this.path}/:*`, authMiddleware)
-      .post(this.path, validationMiddleware(CreateIncomeDto), this.createIncome)
-      .delete(`${this.path}/:id`, this.deleteIncome)
-      .put(`${this.path}/:id`, this.modifyIncome);
+      .post(
+        this.path,
+        authMiddleware,
+        validationMiddleware(CreateIncomeDto),
+        this.createIncome
+      )
+      .delete(`${this.path}/:id`, authMiddleware, this.deleteIncome)
+      .put(`${this.path}/:id`, authMiddleware, this.modifyIncome);
   }
 
-  private getAllIncomes = (req: express.Request, res: express.Response) => {
-    this.income.find().then((incomes) => res.send(incomes));
+  private getAllIncomes = (req: RequestWithUser, res: express.Response) => {
+    this.income
+      .find({ User: req.user?._id })
+      .then((incomes) => res.send(incomes));
   };
 
   private getIncomeById = (
-    req: express.Request,
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
     this.income.findById(req.params.id).then((income) => {
       if (income) {
-        res.send(income);
+        if (income.User === req.user?._id) {
+          res.send(income);
+        } else {
+          next(new NotAuthorizedException());
+        }
       } else {
         next(new IncomeNotFoundException(req.params.id));
       }
@@ -47,7 +60,7 @@ class IncomeController implements Controller {
   };
 
   private modifyIncome = (
-    req: express.Request,
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
@@ -56,33 +69,50 @@ class IncomeController implements Controller {
       .findByIdAndUpdate(req.params.id, incomeData, { new: true })
       .then((income) => {
         if (income) {
-          res.send(income);
+          console.log(income.User);
+          console.log(req.user?._id);
+          if (income.User === req.user?._id) {
+            res.send(income);
+            // !! whyyyyy
+          } else {
+            next(new NotAuthorizedException());
+          }
         } else {
           next(new IncomeNotFoundException(req.params.id));
         }
       });
   };
 
-  private createIncome = (req: express.Request, res: express.Response) => {
-    const incomeData: Income = req.body;
-    const createdIncome = new this.income(incomeData);
-    createdIncome.save().then((savedIncome) => {
-      res.send(savedIncome);
-    });
-  };
-
-  private deleteIncome = (
-    req: express.Request,
+  private createIncome = (
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
-    this.income.findByIdAndDelete(req.params.id).then((successResponse) => {
-      if (successResponse) {
-        res.send(200);
-      } else {
-        next(new IncomeNotFoundException(req.params.id));
-      }
-    });
+    const incomeData: Income = req.body;
+    const createdIncome = new this.income(incomeData);
+    if (req.user) {
+      createdIncome.User = req.user?._id;
+      createdIncome.save().then((savedIncome) => {
+        res.send(savedIncome);
+      });
+    } else next(new NotAuthorizedException());
+  };
+
+  private deleteIncome = async (
+    req: RequestWithUser,
+    res: express.Response,
+    next: express.NextFunction
+  ) => {
+    const incomeToDelete = await this.income.findById(req.params._id);
+    if (incomeToDelete?.User === req.user?._id) {
+      this.income.findByIdAndDelete(req.params.id).then((successResponse) => {
+        if (successResponse) {
+          res.send(200);
+        } else {
+          next(new IncomeNotFoundException(req.params.id));
+        }
+      });
+    } else next(new NotAuthorizedException());
   };
 }
